relay/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-jose[cryptography]==3.3.0
google-auth==2.37.0
requests>=2.31.0
pydantic==2.10.4
pydantic-settings==2.7.1
stripe>=8.0.0
